+++ src/i18n/translator.ts
@@ -1,10 +1,14 @@
-import messagesTranslations from './messages.translations';
+import de from './languages/de.translations';
+import en from './languages/en.translations';
 import { init } from 'i18next';
 
 export const initTranslator = async () => {
     await init({
         lng: 'en',
         fallbackLng: 'en',
-        resources: messagesTranslations,
+        resources: {
+            en,
+            de,
+        },
     });
 };